feat: Permissioned Across-v2 bond token
This PR creates a new bond token, to be used as the collateral required
by the HubPool on the following functions:
 - proposeRootBundle()
 - disputeRootBundle()

The new bond token (symbol "ABT") hosts an internal set of approved
proposer addresses in order to impose constraints on attempted inbound
ABT transfers to the HubPool. Permissioned addresses are unconstrained
in their transfer of ABT, whereas all other addresses must _not_ be the
proposer of a root bundle in order to transfer ABT to the HubPool. This
permits any ABT holder to submit a dispute of an existing root bundle
proposal, whilst preventing non-approved addresses from proposing.

ABT inherits from the WETH9 contract with minimal modifications
necessary to override the transferFrom() function.