chore: Import WETH9 (#227)
Done in advance of the new Across Bond Token (name tbc)
implementation, which will inherit from WETH9.

 - Rename the existing WETH9 -> WETH9Interface
 - Import WETH9 from UMAproject/protocol.
 - WETH9 licensing corrections.

The reason for retaining the WETH9 interface description is that it's
less restrictive than the WETH9 contract that is being imported.
Specifically, Solidity is unhappy that the WETH9 contract identifies
payable functions, and that has ripple effects for the existing users
of the WETH9 interface. We might consider addressing those points,
but the focus here is just to shuffle the furniture to make space for
the new bond token.

Further changes to come once this is in.