improve: [N03] Redundant condition in if statement
## From Audit

In the SpokePool contract, the _updateCountFromFill function is responsible for
updating the fillCounter data when a fill takes place. No update is necessary in the
following three cases: when the action is either a slow fill or an initial zero-fill, or when a partial
fill for that request has already happened.
However, in the case of a zero-fill the execution flow will not reach this function, which makes
the condition endingFillAmount == 0 redundant.
Consider removing the redundant condition for increased clarity, readability, and gas savings.
For extra safety, consider adding a comment where the function returns if there is a zero-fill,
denoting that any updates to this part should take the function _updateCountFromFill into
account.