improve: [L-01] Incorrect or misleading documentation (#251)
Several instances of docstrings or comments in the codebase were found to be erroneous. In
particular:

- In the SpokePool contract's `_fillRelay` function, the comment on line 1003
appears unrelated to the if statement on the following line. Consider revising the
comment.
- The docstring for `pauseDeposits` in the SpokePool contract says it "pauses deposit
and fill functions", but pausing fill functions is performed by calling `pauseFills` , which
does not have a docstring.

Consider revising the `pauseDeposits` docstring and adding a docstring to `pauseFills`.