improve(SpokePool): Apply rounding consistently to partial and full fills
Currently the `amountToSend` could have rounding errors based on the division in `_computeAmountPreFees` and `_computeAmountPostFees`. The rounding errors are applied differently for partial fills and full fills, the latter of which only applies the preFees computation, while the former applies both preFees and postFees.

This PR enforces consistently both the preFees and postFees computations to `amountToSend`