fix: SpokePool deployment corrections (#342)
The Ethereum SpokePool deployment script was unable to correctly resolve
the HubPool address because it assumed that a companionNetworks entry
was defined in the HardhatUserConfig in hardhat.config.ts. Resolve this
by adjusting the Ethereum (HubPool) SpokePool deployment script to avoid
using companionNetworks.

Additionally, the other SpokePool deployment scripts incorrectly logged
the chain ID for the HubPool deployment. This change now factors out the
task of resolving the chain IDs and returns them in an unambiguous form.
This also simplifies most of the deployment scripts.