feat: deposit() wrapper with current quoteTimestamp (#357)
depositNow() is a new deposit() wrapper that removes the need for the
caller to supply a quoteTimestamp. Instead, the value of
getCurrentTime() is used.

This is introduced to address the issue of multisig depositors having
only a narrow window to collect the necessary signatures for a deposit.
Instead, LP fee certainty is traded for flexibility around when the
deposit is able to be submitted. The depositor is still able to query
the expected LP fee via the Across API, so the chance of paying more
fees than anticipated is fairly low.