refactor: Simplify Deposit test + helpers
Simplify deposit generation by converting getDepositParams() to accept
an object, rather than a fixed list of parameters. This makes it cleaner
for the caller to opt-out of supplying various inputs.

This was prompted by a need to update the deposit() and
getDepositParams() utilities such that the caller can supply a custom
message.