improve(SpokePool): Convert wrappedNativeToken from variable to overridable public function
The motivation is to remove an unneccessary SLOAD to read `wrappedNativeToken` on each deposit since this address should never change after a SpokePool is deployed.

Unfortunately, we can't easily convert the existing variable to be an `immutable` variable since upgradeable contracts are not compatible easily with immutable variables [docs](https://docs.openzeppelin.com/upgrades-plugins/1.x/faq#why-cant-i-use-immutable-variables).

This variable is tricky to change to a constant also because it needs to change for each different child `SpokePool` contract implementation.

One way around it is is to rename but not remove the existing `wrappedNativeToken` variable to preserve the storage layout of the contract. Note, I **think** this is safe to do but I'm not sure. Then, declare an overridable `public` function `wrappedNativeToken` that we can reimplement in each child SpokePool to return a constant value. We also could make this function `internal` if that saves more gas versus calling a `public` function