Add x-function lock
Adds a cross-function lock to prevent certain functions from being called atomically (within the same block)

## Gas Cost Analysis

This solution adds an SSTORE and SSLOAD for each fill or execute but it touches only two slots so its not SSTORE-ing into a fresh slot after the first txn and saves gas over the long term.

Adds a new private `mapping` to the PolygonSpokePool with a two slots that are touched: "executeLeaf" and "fill".

For each fill or executeLeaf function:
- Add one 2100 gas SLOAD for the first time the "executeLeaf" or "fill" slot in the new mapping is used, i.e. for the first `fillRelay` or `executeRelayerRefundLeaf` within a block. This SLOAD incurs 100 for every subsequent same-transaction-type within the same block.
- This also adds 5000 gas SSTORE for the first `fill` or `executeRelayerRefundLeaf` within the block (or 22100 for the very first time this slot goes from 0-->N). All other same-transactions should cost one 100 gas SLOAD within the same block instead of the SSTORE since the new mapping is only set once per block for a given transaction type.

So I think the cost additions are:
- multiple or one fill without messages: 0
- first fill with messages or leaf execution: 2100 SLOAD + KECCAK256 cost * 2 + 2100 SLOAD + 5000 SSTORE = 9200 + KECCAK256 * 2
- N fills with messages or N leaf executions:  (9200 + KECCAK256 * 2) + N * (100 SLOAD + KECCAK256 * 2 + 100 SLOAD)

[gas cost source](https://hackmd.io/@fvictorio/gas-costs-after-berlin#Putting-it-all-together)