refactor: Rename fillsRefundedIpfsHash -> fillsRefundedHash (#419)
This is another case of a variable being missed when the naming was
changed. The data availability layer isn't bound to IPFS and will
probably be something else anyway. This is within the scope of the UMIP,
the contracts don't need to be specific about the underlying technology
that is used.