feat: Remove SpokePool v2 functionality (#456)
* feat: Remove non-USS functionality from SpokePool

This PR showcases how we will remove all non-USS functionality in the spoke pools once all non-USS deposits are fills, fills are refunded, and roots are executed.

This PR reduces the bytecode significantly and means we can use the full optimizer runs setting for compiling these new contracts

* Remove Deposit and SlowRelay tests

Signed-off-by: nicholaspai <npai.nyc@gmail.com>

* WIP

* WIP

* Update utils.ts

* Add back Deposit functions

* add back legacy deposit unit tests

* Update hardhat.config.ts

* fix tests

* Update package.json

* fix

Signed-off-by: nicholaspai <npai.nyc@gmail.com>

* Add depositUSSNow, replace error strings with custom error types, fix tests

* Fix

* Fix tests

* Update SpokePool.Deposit.ts

* feat: Remove all "USS" references in contracts

This PR implements a simple find-and-replace of "USS" with "V3" to make it more clear that this is a new version of the Across protocol
that supports new features such as cross-chain token-swaps. USS is an internal code name for "Universal Settlement Service" but this is
not neccessariy the language that will be used to market this product. So, renaming to V3 makes these contract changes more of a blank
slate.

* Replace USS with V3

* Remove UBA references

* fix tests

* merge master

* fix tests

* change depositV3Now to take in exclusivity deadline

* Fix

* Update Ovm_SpokePool.sol

* Update Ovm_SpokePool.sol

* add back depositfor, remove special polygon spoke pool compiler rules

* update hardhat-zksolc

* Update test/SpokePool.Deposit.ts

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>

* Update hardhat.config.ts

---------

Signed-off-by: nicholaspai <npai.nyc@gmail.com>
Co-authored-by: nicholaspai <npai.nyc@gmail.com>
Co-authored-by: nicholaspai <9457025+nicholaspai@users.noreply.github.com>