feat(SpokePool): Add depositExclusive() function to support exclusive order routing
We want to support a user flow where a depositor can be paired off-chain with an `exclusiveRelayer` and give them a very short exclusivity window (think ~2-4 seconds) but we don't want the depositor to have to consider the latency between their deposit txn being submitted to the mempool and when it is mined. Instead, good depositor UX essentially requires that an offset can be specified so that a depositor can specify the amout of exclusivity the filler can get after a deposit is mined.