[WIP] feat: Define in-place alternative to depositId that preserves deposit hash uniqueness while allowing deterministic fill hash computations for honest depositors
## Summary

This PR replaces the contract-wide counter of `depositId` with a per-depositor deposit counter.

## Motivation

The motivation for this PR is to allow honest depositors to broadcast ahead of their deposit transaction to fillers that they want to work with and pre-compute their next deposit hash. This would allow depositors to share with fillers off-chain a Permit2 signed ERC7683 intent + Across deposit and then let the filler submit the deposit txn + immediately fill the user because they would know the precomputed relay hash.

The major benefit of this PR is that re-org risk is completely eliminated for any deposits submitted by "honest" depositors who choose not to try to grief fillers by front-running their fill transactions. This PR accompanied with some sort of off-chain infrastructure where depositors can credibly commit to submitting a deposit (e.g. via an account locking system or by staking collateral) would be the ideal end use case.

A nice property of this PR is that certain properties of Across' order flow system are kept intact:
- deposit ID's are always increasing *for a specific depositor address*
- locating a historical deposit using deposit ID's using a binary search is still efficient assuming you know which depositor you are looking up a deposit for
- All deposits produce unique relay hashes, since the combination of {depositor, depositId} is guaranteed unique

## Gas cost implications

Every first time deposit for a depositor incurs a 20,000 SSTORE cost, and every subsequent deposit for that depositor incurs only a 5,000 SSTORE cost

We assume that most depositors re-use `depositor` addresses so the amortized gas cost is only trivially increased by this PR due to one additional ADD operation