refactor: Relocate message handler interface
The MulticallHandler contract currently imports the SpokePool contract
and experienced a bytecode change after the following unrelated change:

86593c6

This change resulted in its CREATE2 deployment address changing. The
MulticallHandler doesn't actually need any of the SpokePool
implementation or its dependencies, and it seems like we'd like the
MulticallHandler to be deployed to the same address on as many chains as
possible, so factor out the interface to a separate file to insulate it
from any unintended changes.