improve(deploy): Improve SpokePool deployment UX (#577)
The majority of SpokePool deployment scripts have been updated to
support deploying only implementation contracts. This is done with the
knowledge that there are already pre-existing deployments on these
chains, but it introduces a risk that using one of these scripts as the
template for a new chain will inadvertently result in skipping
deployment of the proxy contact. Instead, whether to deploy a new proxy
or not can be inferred by whether the contract exists in
deployments.json.