feat: Use @openzeppelin/contracts@5.x contracts in contracts we can upgrade
Imports an [aliased npm package](https://forum.openzeppelin.com/t/coexist-of-v5-and-v4-contracts/38030/5) so we can use multiple oz libraries in the same solidity contracts.

This way we get access to new and improved contracts. I make changes to all contracts that we can upgrade, such as SpokePools and NOT the HubPool.

For example, I can use the new SafeERC20 method `forceApprove` instead of `safeIncreaseAllowance` which ensures [compatibility with tokens like USDT](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/4231) that make sure all approvals are set to 0 before granting a new approval. This hasn't been an issue so far because we always safeIncreaseAllowance to some number and use the complete allowance, but its worth safety checking.

Other changes:
- Moved `MerkleDistributor` out of uma/core into this repo to reduce dependency on this external repo
- Replaced isCode() call with now [recommended](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/3945) explicit .code.length check
- Explicitly set SpokePoolVerifier pragma to 0.8.19 and removed the overrides in hardhat.config.ts
- Removed unused import in PolygonERC20Test