fix: Support OP bridged USDC in WorldChain adapter (#646)
World Chain implements Circle's bridged USDC standard. Deposits &
withdrawals cannot be initiated via the L1StandardBridge and
L2StandardBridge contracts respectively. Instead they must be sent to
World Chain's own implementations of Circle's bridged USDC standard.

In recognition of the fact that this will likely be a repeating pattern
for new deployments, restyle the World Chain adapter as a generic OP
adapter and instead rely on deployment scripts to configure it with the
correct bridges.

Circle's standard is described here:
https://github.com/circlefin/stablecoin-evm/blob/master/doc/bridged_USDC_standard.md

The process for bridged USDC deposits & withdrawals is here:
https://github.com/defi-wonderland/opUSDC/tree/main?tab=readme-ov-file#l1--l2-usdc-canonical-bridging