feat: Basic SpokePool verification task (#582)
This task intends to automate reading a small number of variables from
the SpokePool on a given chain. This is currently limited to the set of
common variables that are shared amongst all SpokePool deployments, but
could be extended to query chain-specific SpokePool variables.

This should be run by any operator deploying a SpokePool to a new chain,
and may also be integrated into CI.

---------

Signed-off-by: Paul <108695806+pxrl@users.noreply.github.com>