improve(OP): More generic OP chain adapter deployment
This change permits an incremental step towards a totally generic chain
adapter deployment for OP stack chains. It does this be restructuring
the address mappings for OP stack chains, such that the L1 contract
addresses can be resolved solely with the combination of HubPool and
SpokPool chain IDs. It also updates the new OP_Adapter deployment script
to support a runtime-defined SpokePool chain ID. This is clunkily
supplied via the environment because hardhat deploy doesn't seem to
permit arbitrary runtime arguments being supplied, but that aside, this
should ease subsequent deployments.