improve: use immutable `SHARED_BRIDGE` whenever possible (#748)
In the `ZkStack_*_Adapte`, we assume that the bridge hub's shared bridge is immutable. While this is not true -- the shared bridge may be changed in the future -- it is assumed to be immutable so that L1 calls will be cheaper, as we stand to save ~2 staticcalls per message. The hope is in the long run, if these bridges change, the cost of redeploying an adapter will have already been "paid off" by the gas savings of assuming immutability.

The issue `L-03` proposed by Openzeppelin proposed the opposite solution: do not make `SHARED_BRIDGE` an immutable variable; however, they did also approve of using an immutable variable under these circumstances. 
Signed-off-by: bennett <bennett@umaproject.org>