improve(tasks): Auto deposit-and-burn for new LP tokens
When adding new deposit routes, if the LP token for the given L1 token
does not yet exist, create the LP token and atomically deposit and burn
a single unit. This prevents the HubPool from ever having a 0 balance of
the l1 token, protecting against a known issue.

It's been on the todo list to automate this for a while, but it hasn't
been prioritised because adding LP tokens is quite infrequent.
Atomic deposit-and-burn was done manually last time
(0x566087fbaa74a32ec94bf4ba58c1451bf916991b686c0895245f726339b1725d),
but it's preferable to automate this as much as possible to remove the
human element.