feat: Support USDC on Lens (#986)
Lens uses Circle's Bridged (upgradable) USDC standard and requires a custom bridge for it. On mainnet, the chain adapter contract needs a modified workflow to make deposits. On the L2 it's as simple as depositing via a different contract address, albeit with the same contract interface as the standard ERC20 bridge.

As part of this change, both ends are updated for eventual CCTP compatibility. This introduces some configuration nuance to the constructor arguments of both contracts:

usdcAddress = 0x0 => USDC tokens are routed via the standard ERC20 bridge (current required for zkSync USDC.e).
usdcAddress != 0x0 AND cctpTokenMessenger != 0x0 => USDC tokens are routed via Circle's CCTP bridges.
usdcAddress != 0x0 AND zkUSDCBridge != 0x0 => USDC tokens are routed via the custom bridge for Circle Bridged (upgradable) USDC.
A simple check on the various constructor arguments imposed during deployment, but it's highly recommended to manually verify the configurations post-deployment.

Implementation according to the Matter Labs guidance at https://github.com/matter-labs/usdc-bridge

This is immediately for Lens but is generalised to apply to any zkSync-stack deployment.

This implementation was audited by OpenZeppelin.

Signed-off-by: Reinis Martinsons <reinis@umaproject.org>