feat: Create PermissionedMulticall3
We're running into a frustrating situation where a contract like `SP1Helios.update()` can only be called by a permissioned EOA but we'd like to batch this call with an unpermissioned call to `Universal_SpokePool.execute()`.

This doesn't work because there is no way to batch the `update()` call from an EOA with a contract call.

Therefore if we had a function like this PermissionedMulticall3 that we could whitelist as a `SP1Helios.update()` caller, then we could batch it with `execute()`