improve(tasks): Auto deposit for new LP tokens (#901)
When adding new deposit routes, if the LP token for the given L1 token
does not yet exist, create the LP token and atomically deposit and burn
a single unit. This prevents the HubPool from ever having a 0 balance of
the l1 token, protecting against a known issue.

It's been on the todo list to automate this for a while, but it hasn't
been prioritised because adding LP tokens is quite infrequent.

While here, also add protection against setting deposit routes for v4
chains.