L-01 deposit Will Not Work for Non-EVM Target Chains
The deposit function of the SpokePoolPeriphery contract allows users to deposit native value to the SpokePool.

However, its recipient and exclusiveRelayer arguments are both of type address and are then casted to bytes32. As a result, it is not possible to bridge wrapped native tokens to non-EVM blockchains.

Consider changing the type of the recipient and exclusiveRelayer arguments of the deposit function so that callers are allowed to specify non-EVM addresses for deposits.