Rename deposit function to depositNative for clarity (#1019)
The deposit function of the SpokePoolPeriphery contract allows users to deposit native value to the SpokePool.

Although it is possible to specify the inputToken parameter, it is not possible to deposit other tokens through this function. Because of that, it could be renamed to depositNative or a similar name in order to make this fact clear.

Consider renaming the deposit function in order to improve readability of the codebase.