chore: Drop Linea SDK dependency
The Linea SDK is a painful dependency in that it depends on
better-sqlite3, and that seems to have build issues when migrating to
node 22.

The Linea SDK is only required here for a single script that's never
actually been needed in production. To simplify things, we can migrate
it to the relayer repository where we already have to deal with the
Linea SDK dependencies.