chore: Bump node 16 (20) -> 22 (#1065)
* chore: Drop Linea SDK dependency

The Linea SDK is a painful dependency in that it depends on
better-sqlite3, and that seems to have build issues when migrating to
node 22.

The Linea SDK is only required here for a single script that's never
actually been needed in production. To simplify things, we can migrate
it to the relayer repository where we already have to deal with the
Linea SDK dependencies.

* chore: Bump node 16 -> 22

16 is long EoL but we were in practice using 20 in CI.

* Tighten version check

* Simplify check

* chore: Bump kit 2.1.0 -> 2.3.0

* Revert "chore: Bump kit 2.1.0 -> 2.3.0"

This reverts commit 73b6e4fabdedf49f696d304cc90c62dbd78758ce.

* Bump to 22.18

* Bump node

* Revert "Bump node"

This reverts commit 121e6750cc118d06a2c6f7595d6800e8df98510f.